Fix build failure on windows runners. (#251)
build-tools doesn't need to be a dependency for the plugin as plugin config is referencing the files directly using absolute path. This is causing problem when the underlying .m2 directory is ephemeral between actions steps and the installed jar is lost which causes the next step (clean install) to fail.
Also removing `mvn test` step as per maven lifecycle previous `mvn install` already runs the `test` phase.